double to float trunc warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8911 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/geometry/TankSceneNode.cxx
@@ -397,7 +397,7 @@ void TankSceneNode::setJumpJets(float scale)
       jumpJetsGroundLights[i].setPosition(jetPos);
 
       // setup the random lengths
-      const float randomFactor = (1.0f - (0.5f * (0.5f - bzfrand())));
+      const float randomFactor = (1.0f - (0.5f * (0.5f - (float)bzfrand())));
       jumpJetsLengths[i] = jumpJetsScale * randomFactor;
     }
   }
